requirements.txt
fastapi
uvicorn
python-multipart
supabase
opencv-python-headless
imagehash
google-genai
requests
python-dotenv
Pillow
